PR-10 Move database directory
