+++ app/src/main/java/com/pbs/expenseApp/database/repositories/Category.kt
@@ -1,6 +1,6 @@
-package com.pbs.expenseApp.ui.database.repositories
+package com.pbs.expenseApp.database.repositories
 
-import com.pbs.expenseApp.ui.database.entities.Category
+import com.pbs.expenseApp.database.entities.Category
 import kotlinx.coroutines.flow.Flow
 
 interface CategoryRepository {
