+++ app/src/main/java/com/pbs/expenseApp/database/repositories/Expense.kt
@@ -1,6 +1,6 @@
-package com.pbs.expenseApp.ui.database.repositories
+package com.pbs.expenseApp.database.repositories
 
-import com.pbs.expenseApp.ui.database.entities.Expense
+import com.pbs.expenseApp.database.entities.Expense
 import kotlinx.coroutines.flow.Flow
 
 interface ExpenseRepository {
