+++ app/src/main/java/com/pbs/expenseApp/ui/screens/user/UserEntryViewModel.kt
@@ -5,8 +5,8 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.pbs.expenseApp.ui.database.entities.User
-import com.pbs.expenseApp.ui.database.repositories.UserRepository
+import com.pbs.expenseApp.database.entities.User
+import com.pbs.expenseApp.database.repositories.UserRepository
 import kotlinx.coroutines.async
 import kotlinx.coroutines.launch
